+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Cover images weren't getting inserted into brand new notes
 - Prevent the note list from flashing while typing on non-compact layouts
 - Note UI didn't show recently entered text on background/restore
+- If you delete an image from a note, it disappears from the note list
 
 ## [0.47.0] - 2021-05-22